fixed problem with CORS in spring security
