+++ Backend/src/main/java/com/homeappliancesshop/config/CorsConfig.java
@@ -1,0 +1,18 @@
+package com.homeappliancesshop.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class CorsConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:3000") // Dodaj tutaj adres URL Twojego klienta
+                .allowedMethods("GET", "POST", "PUT", "DELETE") // Dodaj tutaj obsługiwane metody HTTP
+                .allowedHeaders("*")
+                .allowCredentials(true);
+    }
+}
